+++ FirebaseAuth/Sources/AuthProvider/GameCenter/FIRGameCenterAuthProvider.m
@@ -54,37 +54,70 @@ NS_ASSUME_NONNULL_BEGIN
     }
     return;
   }
-
-  [localPlayer generateIdentityVerificationSignatureWithCompletionHandler:^(
-                   NSURL *publicKeyURL, NSData *signature, NSData *salt, uint64_t timestamp,
-                   NSError *error) {
-    if (error) {
-      if (completion) {
-        completion(nil, error);
-      }
-    } else {
-      if (completion) {
-        /**
-         @c `localPlayer.alias` is actually the displayname needed, instead of
-         `localPlayer.displayname`. For more information, check
-         https://developer.apple.com/documentation/gamekit/gkplayer
-         **/
-        NSString *displayName = localPlayer.alias;
-// iOS 13 deprecation
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+  if (@available(iOS 13.5, macOS 10.15.5, macCatalyst 13.5,
+  // Availability of fetchItemsForIdentityVerificationSignature was extended to tvOS 13.4.8 in
+  // Xcode 14 (from tvOS 13.5 in Xcode 13).
+  // TODO: Remove this check when we no longer support Xcode 13.
+#if defined(__TVOS_16_0)
+                 tvOS 13.4.8,
+#else
+                 tvOS 13.5,
+#endif
+                     *)) {
+    [localPlayer fetchItemsForIdentityVerificationSignature:^(
+                     NSURL *_Nullable publicKeyURL, NSData *_Nullable signature,
+                     NSData *_Nullable salt, uint64_t timestamp, NSError *_Nullable error) {
+      if (error) {
+        if (completion) {
+          completion(nil, error);
+        }
+      } else {
         FIRGameCenterAuthCredential *credential =
             [[FIRGameCenterAuthCredential alloc] initWithPlayerID:localPlayer.playerID
+                                                     teamPlayerID:localPlayer.teamPlayerID
+                                                     gamePlayerID:localPlayer.gamePlayerID
                                                      publicKeyURL:publicKeyURL
                                                         signature:signature
                                                              salt:salt
                                                         timestamp:timestamp
-                                                      displayName:displayName];
-#pragma clang diagnostic pop
+                                                      displayName:localPlayer.displayName];
         completion(credential, nil);
       }
-    }
-  }];
+    }];
+  } else {
+    [localPlayer generateIdentityVerificationSignatureWithCompletionHandler:^(
+                     NSURL *publicKeyURL, NSData *signature, NSData *salt, uint64_t timestamp,
+                     NSError *error) {
+      if (error) {
+        if (completion) {
+          completion(nil, error);
+        }
+      } else {
+        if (completion) {
+          /**
+           @c `localPlayer.alias` is actually the displayname needed, instead of
+           `localPlayer.displayname`. For more information, check
+           https://developer.apple.com/documentation/gamekit/gkplayer
+           **/
+          NSString *displayName = localPlayer.alias;
+// iOS 13 deprecation
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+          FIRGameCenterAuthCredential *credential =
+              [[FIRGameCenterAuthCredential alloc] initWithPlayerID:localPlayer.playerID
+                                                       teamPlayerID:nil
+                                                       gamePlayerID:nil
+                                                       publicKeyURL:publicKeyURL
+                                                          signature:signature
+                                                               salt:salt
+                                                          timestamp:timestamp
+                                                        displayName:displayName];
+#pragma clang diagnostic pop
+          completion(credential, nil);
+        }
+      }
+    }];
+  }
 }
 
 @end