+++ FirebaseInstallations/Source/Tests/Utils/FIRTestKeychain.m
@@ -36,8 +36,11 @@
         return nil;
       }
     }
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     OSStatus result = SecKeychainCreate([keychainPath cStringUsingEncoding:NSUTF8StringEncoding], 0,
                                         "1", false, nil, &privateKeychain);
+#pragma clang diagnostic pop
     if (result != errSecSuccess) {
       NSError *error = [NSError errorWithDomain:NSOSStatusErrorDomain code:result userInfo:nil];
       NSLog(@"SecKeychainCreate error: %@", error);
@@ -50,7 +53,10 @@
 
 - (void)dealloc {
   if (self.testKeychainRef) {
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     OSStatus result = SecKeychainDelete(self.testKeychainRef);
+#pragma clang diagnostic pop
     if (result != errSecSuccess) {
       NSError *error = [NSError errorWithDomain:NSOSStatusErrorDomain code:result userInfo:nil];
       NSLog(@"SecKeychainCreate error: %@", error);