+++ FirebaseRemoteConfig/Sources/RCNConfigSettings.m
@@ -110,6 +110,11 @@ static const int kRCNExponentialBackoffMaximumInterval = 60 * 60 * 4;  // 4 hour
     }
 
     _isFetchInProgress = NO;
+    _lastTemplateVersion = [_userDefaultsManager lastTemplateVersion];
+    _realtimeExponentialBackoffRetryInterval =
+        [_userDefaultsManager currentRealtimeThrottlingRetryIntervalSeconds];
+    _realtimeExponentialBackoffThrottleEndTime = [_userDefaultsManager realtimeThrottleEndTime];
+    _realtimeRetryCount = [_userDefaultsManager realtimeRetryCount];
   }
   return self;
 }
@@ -233,7 +238,51 @@ static const int kRCNExponentialBackoffMaximumInterval = 60 * 60 * 4;  // 4 hour
       [[NSDate date] timeIntervalSince1970] + randomizedRetryInterval;
 }
 
-- (void)updateMetadataWithFetchSuccessStatus:(BOOL)fetchSuccess {
+/// If the last Realtime stream attempt was not successful, update the (exponential backoff) period
+/// that we wait until trying again. Any subsequent Realtime requests will be checked and allowed
+/// only if past this throttle end time.
+- (void)updateRealtimeExponentialBackoffTime {
+  // If there was only one stream attempt before, reset the retry interval.
+  if (_realtimeRetryCount == 0) {
+    FIRLogDebug(kFIRLoggerRemoteConfig, @"I-RCN000058",
+                @"Throttling: Entering exponential Realtime backoff mode.");
+    _realtimeExponentialBackoffRetryInterval = kRCNExponentialBackoffMinimumInterval;
+  } else {
+    FIRLogDebug(kFIRLoggerRemoteConfig, @"I-RCN000058",
+                @"Throttling: Updating Realtime throttling interval.");
+    // Double the retry interval until we hit the truncated exponential backoff. More info here:
+    // https://cloud.google.com/storage/docs/exponential-backoff
+    _realtimeExponentialBackoffRetryInterval =
+        ((_realtimeExponentialBackoffRetryInterval * 2) < kRCNExponentialBackoffMaximumInterval)
+            ? _realtimeExponentialBackoffRetryInterval * 2
+            : _realtimeExponentialBackoffRetryInterval;
+  }
+
+  // Randomize the next retry interval.
+  int randomPlusMinusInterval = ((arc4random() % 2) == 0) ? -1 : 1;
+  NSTimeInterval randomizedRetryInterval =
+      _realtimeExponentialBackoffRetryInterval +
+      (0.5 * _realtimeExponentialBackoffRetryInterval * randomPlusMinusInterval);
+  _realtimeExponentialBackoffThrottleEndTime =
+      [[NSDate date] timeIntervalSince1970] + randomizedRetryInterval;
+
+  [_userDefaultsManager setRealtimeThrottleEndTime:_realtimeExponentialBackoffThrottleEndTime];
+  [_userDefaultsManager
+      setCurrentRealtimeThrottlingRetryIntervalSeconds:_realtimeExponentialBackoffRetryInterval];
+}
+
+- (void)setRealtimeRetryCount:(int)realtimeRetryCount {
+  _realtimeRetryCount = realtimeRetryCount;
+  [_userDefaultsManager setRealtimeRetryCount:_realtimeRetryCount];
+}
+
+- (NSTimeInterval)getRealtimeBackoffInterval {
+  NSTimeInterval now = [[NSDate date] timeIntervalSince1970];
+  return _realtimeExponentialBackoffThrottleEndTime - now;
+}
+
+- (void)updateMetadataWithFetchSuccessStatus:(BOOL)fetchSuccess
+                             templateVersion:(NSString *)templateVersion {
   FIRLogDebug(kFIRLoggerRemoteConfig, @"I-RCN000056", @"Updating metadata with fetch result.");
   [self updateFetchTimeWithSuccessFetch:fetchSuccess];
   _lastFetchStatus =
@@ -243,6 +292,7 @@ static const int kRCNExponentialBackoffMaximumInterval = 60 * 60 * 4;  // 4 hour
     [self updateLastFetchTimeInterval:[[NSDate date] timeIntervalSince1970]];
     // Note: We expect the googleAppID to always be available.
     _deviceContext = FIRRemoteConfigDeviceContextWithProjectIdentifier(_googleAppID);
+    [_userDefaultsManager setLastTemplateVersion:templateVersion];
   }
 
   [self updateMetadataTable];