+++ FirebaseSessions/Sources/Settings/SettingsProvider.swift
@@ -22,4 +22,14 @@ protocol SettingsProvider {
 
   // API to check if the settings are stale
   func isSettingsStale() -> Bool
+
+  // Config to show if sessions is enabled
+  var sessionsEnabled: Bool? { get }
+
+  // Config showing the sampling rate for sessions
+
+  var samplingRate: Double? { get }
+
+  // Background timeout config value before which a new session is generated
+  var sessionTimeout: TimeInterval? { get }
 }