+++ FirebaseSessions/Tests/Unit/InitiatorTests.swift
@@ -15,53 +15,42 @@
 import XCTest
 @testable import FirebaseSessions
 
-#if os(iOS) || os(tvOS)
-  import UIKit
-#elseif os(macOS)
-  import Cocoa
-  import AppKit
-#elseif os(watchOS)
-  import WatchKit
-#endif
-
 class InitiatorTests: XCTestCase {
   // 2021-11-01 @ 00:00:00 (EST)
   let date = Date(timeIntervalSince1970: 1_635_739_200)
+  let validSettings: [String: Any] = [:]
 
-  func postBackgroundedNotification() {
-    let notificationCenter = NotificationCenter.default
-    #if os(iOS) || os(tvOS)
-      notificationCenter.post(name: UIApplication.didEnterBackgroundNotification, object: nil)
-    #elseif os(macOS)
-      notificationCenter.post(name: NSApplication.didResignActiveNotification, object: nil)
-    #elseif os(watchOS)
-      if #available(watchOSApplicationExtension 7.0, *) {
-        notificationCenter.post(
-          name: WKExtension.applicationDidEnterBackgroundNotification,
-          object: nil
-        )
-      }
-    #endif
-  }
+  var cache: SettingsCacheClient!
+  var appInfo: MockApplicationInfo!
+  var downloader: MockSettingsDownloader!
+  var remoteSettings: RemoteSettings!
+  var localOverrideSettings: LocalOverrideSettings!
+  var sdkDefaultSettings: SDKDefaultSettings!
+  var sessionSettings: SessionsSettings!
 
-  func postForegroundedNotification() {
-    let notificationCenter = NotificationCenter.default
-    #if os(iOS) || os(tvOS)
-      notificationCenter.post(name: UIApplication.didBecomeActiveNotification, object: nil)
-    #elseif os(macOS)
-      notificationCenter.post(name: NSApplication.didBecomeActiveNotification, object: nil)
-    #elseif os(watchOS)
-      if #available(watchOSApplicationExtension 7.0, *) {
-        notificationCenter.post(
-          name: WKExtension.applicationDidBecomeActiveNotification,
-          object: nil
-        )
-      }
-    #endif
+  override func setUp() {
+    super.setUp()
+    appInfo = MockApplicationInfo()
+    cache = SettingsCache()
+    cache.removeCache() // just reinstantiating cache isn't enough because of persistence
+    downloader = MockSettingsDownloader(successResponse: validSettings)
+    remoteSettings = RemoteSettings(appInfo: appInfo, downloader: downloader, cache: cache)
+    remoteSettings.updateSettings(currentTime: Date())
+
+    localOverrideSettings = LocalOverrideSettings()
+    sdkDefaultSettings = SDKDefaultSettings()
+
+    sessionSettings = SessionsSettings(
+      appInfo: appInfo,
+      installations: MockInstallationsProtocol(),
+      sdkDefaults: sdkDefaultSettings,
+      localOverrides: localOverrideSettings,
+      remoteSettings: remoteSettings
+    )
   }
 
   func test_beginListening_initiatesColdStart() throws {
-    let initiator = SessionInitiator()
+    let initiator = SessionInitiator(settings: sessionSettings)
     var initiateCalled = false
     initiator.beginListening {
       initiateCalled = true
@@ -72,7 +61,10 @@ class InitiatorTests: XCTestCase {
   func test_appForegrounded_initiatesNewSession() throws {
     // Given
     var pausedClock = date
-    let initiator = SessionInitiator(currentTimeProvider: { pausedClock })
+    let initiator = SessionInitiator(
+      settings: sessionSettings,
+      currentTimeProvider: { pausedClock }
+    )
     var sessionCount = 0
     initiator.beginListening {
       sessionCount += 1