+++ Firestore/core/src/local/remote_document_cache.h
@@ -19,7 +19,9 @@
 
 #include <string>
 
+#include "Firestore/core/src/model/document_key.h"
 #include "Firestore/core/src/model/model_fwd.h"
+#include "Firestore/core/src/model/overlay.h"
 
 namespace firebase {
 namespace firestore {
@@ -100,17 +102,20 @@ class RemoteDocumentCache {
    *
    * Cached DeletedDocument entries have no bearing on query results.
    *
-   * @param path The collection path to match documents against.
+   * @param query The query to match documents against.
    * @param offset The read time and document key to start scanning at
    * (exclusive).
    * @param limit The maximum number of results to return.
    * If the limit is not defined, returns all matching documents.
+   * @param mutated_docs The documents with local mutations, they are read
+   * regardless if the remote version matches the given query.
    * @return The set of matching documents.
    */
-  virtual model::MutableDocumentMap GetAll(
-      const model::ResourcePath& path,
+  virtual model::MutableDocumentMap GetDocumentsMatchingQuery(
+      const core::Query& query,
       const model::IndexOffset& offset,
-      absl::optional<size_t> limit = absl::nullopt) const = 0;
+      absl::optional<size_t> limit = absl::nullopt,
+      const model::OverlayByDocumentKeyMap& mutated_docs = {}) const = 0;
 
   /**
    * Sets the index manager used by remote document cache.